Display lamp status only once at the start of each request
Elevator should trigger the floor lamp status signal to floorsubsystem only when the request is taken and it has decided whether it’s going up or down based on the recieved request between x and y floor. It should not trigger (and thus print the lamp status in the console) every time the elevator moves a floor.