Added Scheduler States + corrected impacted tests
AwaitingRequestState: The Scheduler is idle and waiting for new elevator requests from the FloorSubsystem.
RequestProcessingState: The Scheduler is processing a new elevator request, selecting the appropriate elevator, and preparing to send the request to the ElevatorSubsystem.
ElevatorDispatchState: The Scheduler is sending the elevator request to the selected elevator in the ElevatorSubsystem.

Tests: Removed the test addToResponseLog in Schedulertest.java as its not needed and updated the testSelectElevator test.